+++ index.html
@@ -28,9 +28,9 @@
 
           <nav>
             <ul>
-              <li><a href="#">ALL PRODUCTS</a></li>
-              <li><a href="#">BEST</a></li>
-              <li><a href="#">NEW</a></li>
+              <li><a href="#" data-navigo="/product/all">ALL PRODUCTS</a></li>
+              <li><a href="#" data-navigo="/product/best">BEST</a></li>
+              <li><a href="#" data-navigo="/product/new">NEW</a></li>
             </ul>
           </nav>
 
@@ -40,47 +40,20 @@
               <button class="material-symbols-outlined">search</button>
             </form>
 
-            <span class="material-symbols-outlined cart"> local_mall </span>
+            <a href="#" data-navigo="/product/cart">
+              <span class="material-symbols-outlined cart"> local_mall </span>
+            </a>
 
-            <span class="profile material-symbols-outlined">
-              account_circle
-            </span>
+            <a href="#" data-navigo="/login">
+              <span class="profile material-symbols-outlined">
+                account_circle
+              </span>
+            </a>
           </div>
         </div>
       </header>
 
       <div id="app"></div>
-<!--       
-      <div class="visual">
-        <div class="swiper">
-          <div class="swiper-wrapper">
-            <div class="swiper-slide"></div>
-            <div class="swiper-slide"></div>
-            <div class="swiper-slide"></div>
-            <div class="swiper-slide"></div>
-          </div>
-          <div class="swiper-pagination"></div>
-        </div>
-      </div>
-
-      <section>
-        <div class="new-items">
-          <h2>새로 나왔어요</h2>
-
-          <div class="swiper ">
-            <div class="swiper-wrapper">
-            </div>
-            <div class="swiper-button-prev"></div>
-            <div class="swiper-button-next"></div>
-          </div>
-          <span class="loading">loading</span>
-        </div>
-
-        <div class="recommend">
-          <h2>이런 가구 어때요?</h2>
-          <span class="loading">loading</span>
-        </div>
-      </section> -->
 
       <footer>
         <div class="logo">
